Extract taxonomy parsing outside cite page creator
